Fix : Error 모두 수정, Style : hover 디테일 수정, Refactoring : useContentStore 수정
