+++ netflix-assignment/src/app/(route)/layout.tsx
@@ -7,11 +7,9 @@ export default function RouteLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <body>
-        <div className="relative w-[375px] h-[100vh]">
-            {children}
-            <Footer/>
-        </div>
-    </body>
+    <div className="relative w-[375px] h-[100vh]">
+        {children}
+        <Footer/>
+    </div>
   );
 }
